+++ keepalived/check/check_http.c
@@ -24,6 +24,7 @@
 #include "config.h"
 
 #include <openssl/err.h>
+#include <stdint.h>
 #include "check_http.h"
 #include "check_ssl.h"
 #include "check_api.h"
@@ -465,7 +466,7 @@ http_handle_response(thread_t * thread, unsigned char digest[16]
 
 	/* Report a length mismatch the first time we get the specific difference */
 	url = list_element(http_get_check->url, http_get_check->url_it);
-	if (req->content_len != UINT_MAX && req->content_len != req->rx_bytes) {
+	if (req->content_len != SIZE_MAX && req->content_len != req->rx_bytes) {
 		if (url->len_mismatch != (ssize_t)req->content_len - (ssize_t)req->rx_bytes) {
 			log_message(LOG_INFO, "http_check for RS %s VS %s url %s%s: content_length (%lu) does not match received bytes (%lu)",
 				    FMT_RS(checker->rs, checker->vs), FMT_VS(checker->vs), url->virtualhost ? url->virtualhost : "",
@@ -530,17 +531,17 @@ http_process_response(request_t *req, size_t r, bool do_md5)
 			req->content_len = extract_content_length(req->buffer, req->len);
 			r = req->len - (size_t)(req->extracted - req->buffer);
 			if (r && do_md5) {
-				if (req->content_len == UINT_MAX || req->content_len > req->rx_bytes)
+				if (req->content_len == SIZE_MAX || req->content_len > req->rx_bytes)
 					MD5_Update(&req->context, req->extracted,
-						   req->content_len == UINT_MAX || req->content_len >= req->rx_bytes + r ? r : req->content_len - req->rx_bytes);
+						   req->content_len == SIZE_MAX || req->content_len >= req->rx_bytes + r ? r : req->content_len - req->rx_bytes);
 			}
 			req->rx_bytes = r;
 			req->len = 0;
 		}
 	} else if (req->len) {
-		if (req->content_len == UINT_MAX || req->content_len > req->rx_bytes) {
+		if (req->content_len == SIZE_MAX || req->content_len > req->rx_bytes) {
 			MD5_Update(&req->context, req->buffer,
-				   req->content_len == UINT_MAX || req->content_len >= req->rx_bytes + req->len ? req->len : req->content_len - req->rx_bytes);
+				   req->content_len == SIZE_MAX || req->content_len >= req->rx_bytes + req->len ? req->len : req->content_len - req->rx_bytes);
 		}
 		req->rx_bytes += req->len;
 		req->len = 0;
